refactor(SettingsDialog): consolidate files
