+++ client/src/dialogs/DialogRoot.jsx
@@ -8,7 +8,7 @@ import AnalyticsDialog from './Analytics'
 import FeatureFlagDialog from './FeatureFlagDialog'
 import GeotagDialog from './Geotag'
 import SaveAsImageDialog from './SaveAsImage'
-import SettingsDialog from './SettingsDialog'
+import SettingsDialog from './Settings'
 import SignInDialog from './SignInDialog'
 import WhatsNewDialog from './WhatsNewDialog'
 import NewsletterDialog from './NewsletterDialog'
