+++ client/src/dialogs/Settings/SettingsDialog.scss
